Initial Commit
So far, there is the ability to search for recipes with keywords and view nutritional info of them. Hovering over rows will show images of the dish. Clicking on the link will open the recipe page. Next steps is sorting the table based on specific macro, having a list of ingredients in each recipe on a separate page, and favoriting recipes to be stored in a database.